Revert "Revert "Get rid of sdk internal reloadutil package for now. (#9515)" (#9521)"
This reverts commit f687267c8086b63ee10043b42918ea0a63956c39.
